+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # Changelog
 
-## 0.0.1
-- Initial release
+## 0.0.3
+- Perform SNOW CI Lookup during init
+- Perform SNOW User lookup during selection
+- Allow requiring a User when sending to SNOW
+- Default include in Call Controls for RoomOS mode
+- Fix validate Protocol exists when processing call
+- Improve error handling
+- Remove Debug Services button (retain Debug Survey)
 
 ## 0.0.2
 - Capture default submit from survey rating prompt
 - Don't clear call details if Report Issue is submitted during an active call
 
+## 0.0.1
+- Initial release